fix precision level for perf counter reading and add common error hint

+++ perf_profiler.h
@@ -59,7 +59,7 @@ class PerfProfiler {
 
     void Disable();
 
-    std::map<std::string, int> Read();
+    std::map<std::string, uint64_t> Read();
 
 };
